+++ clubs/urls.py
@@ -1,6 +1,6 @@
 '''
-MVP demo ver 0.0.1
-2024.07.22
+MVP demo ver 0.0.2
+2024.07.28
 clubs/urls.py
 
 역할: clubs 앱 내의 URL API 엔드포인트 설정
@@ -8,10 +8,12 @@ clubs/urls.py
 '''
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import ClubViewSet
+from .views import ClubViewSet, ClubAdminViewSet, ClubMemberViewSet
 
 router = DefaultRouter()
 router.register(r'', ClubViewSet)
+router.register(r'', ClubAdminViewSet, basename='club-admin')
+router.register(r'', ClubMemberViewSet, basename='club-member')
 
 urlpatterns = [
     path('', include(router.urls)),
